Financial-Q&A-RAG/requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
langchain-huggingface
docling
faiss-cpu
python-dotenv